# Conflicts:
#	src/main/java/com/mywebapp/controllers/user/BookingInfoController.java
#	src/main/webapp/jsp/service/bookingInfo.jsp

+++ src/main/java/com/mywebapp/controllers/user/RoomDetailController.java
@@ -26,7 +26,7 @@ public class RoomDetailController extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 	    // guestMain.jsp에서 방 id 파라미터 가져오기
-        int roomId = Integer.parseInt(req.getParameter("roomId"));
+        long roomId = Long.parseLong(req.getParameter("roomId"));
         
         System.out.println(roomId);
         // DAO를 사용하여 방의 상세 정보를 가져옵니다.
